added nginx-auth role
