+++ ansible/roles/nginx-auth/defaults/main.yml
@@ -1,0 +1,2 @@
+nginx_user: nginx_user
+nginx_password: nginx_password
