+++ ansible/roles/nginx-auth/files/polkadot_metrics.nginx.conf
