+++ ansible/roles/nginx-auth/molecule/default/Dockerfile.j2
@@ -1,0 +1,27 @@
+# Molecule managed
+
+{% if item.registry is defined %}
+  FROM {{ item.registry.url }}/{{ item.image }}
+{% else %}
+  FROM {{ item.image }}
+{% endif %}
+
+{% if item.env is defined %}
+  {% for var, value in item.env.items() %}
+    {% if value %}
+      ENV {{ var }} {{ value }}
+    {% endif %}
+  {% endfor %}
+{% endif %}
+
+RUN if [ $(command -v apt-get) ]; then apt-get update && apt-get install -y python sudo bash ca-certificates iproute2 wget nginx && apt-get clean; \
+elif [ $(command -v dnf) ]; then dnf makecache && dnf --assumeyes install python sudo python-devel python*-dnf bash iproute && dnf clean all; \
+elif [ $(command -v yum) ]; then yum makecache fast && yum install -y python sudo yum-plugin-ovl bash iproute && sed -i 's/plugins=0/plugins=1/g' /etc/yum.conf && yum clean all; \
+elif [ $(command -v zypper) ]; then zypper refresh && zypper install -y python sudo bash python-xml iproute2 && zypper clean -a; \
+elif [ $(command -v apk) ]; then apk update && apk add --no-cache python sudo bash ca-certificates; \
+elif [ $(command -v xbps-install) ]; then xbps-install -Syu && xbps-install -y python sudo bash ca-certificates iproute2 && xbps-remove -O; fi
+
+RUN wget https://raw.githubusercontent.com/gdraheim/docker-systemctl-replacement/master/files/docker/systemctl.py -O /usr/bin/systemctl && \
+chmod a+x /usr/bin/systemctl && \
+test -L /bin/systemctl || ln -sf /usr/bin/systemctl /bin/systemctl && \
+rm -f /sbin/init && touch /sbin/systemd && ln -sf /sbin/systemd /sbin/init
