+++ ansible/roles/nginx-auth/molecule/default/converge.yml
@@ -1,0 +1,5 @@
+---
+- name: Converge
+  hosts: all
+  roles:
+    - role: nginx-auth
