+++ ansible/roles/node-exporter/tasks/main.yml
@@ -1,52 +1,3 @@
-- name: check if .htpasswd file already exists
-  stat:
-    path: /etc/nginx/.htpasswd
-  register: htpasswd_file
-
-- name: remove libapr related packages
-  apt:
-    pkg:
-    - libapr1
-    - libaprutil1
-    state: absent
-    update_cache: yes
-    autoremove: yes
-  when: not htpasswd_file.stat.exists
-
-- name: install apache2-utils
-  apt:
-    name: apache2-utils
-    state: present
-    update_cache: yes
-  when: not htpasswd_file.stat.exists
-
-- name: generate .htpasswd file
-  shell: |
-    set -o pipefail
-    htpasswd -bc /etc/nginx/.htpasswd {{ node_exporter_user }} {{ node_exporter_password }}
-    chmod 600 /etc/nginx/.htpasswd
-    chown www-data:www-data /etc/nginx/.htpasswd
-  args:
-    executable: /bin/bash
-
-- name: create nginx config for node exporter
-  copy:
-    src: nginx.conf
-    dest: /etc/nginx/sites-enabled/node-exporter.conf
-
-- name: create nginx config for polkadot metrics
-  copy:
-    src: polkadot.nginx.conf
-    dest: /etc/nginx/sites-enabled/polkadot-metrics.conf 
-
-- name: restart nginx service
-  systemd:
-    name: nginx
-    state: restarted
-    daemon_reload: yes
-    enabled: yes
-  changed_when: true
-
 - name: create node_exporter download dir
   file:
     path: /root/node_exporter
