+++ src/lib/clients/ansible.js
@@ -3,7 +3,7 @@ const path = require('path');
 const cmd = require('../cmd');
 const { Project } = require('../project');
 const tpl = require('../tpl');
-const { nodeExporterUsername, nodeExporterPassword } = require('../env');
+const { nginxUsername, nginxPassword } = require('../env');
 
 const inventoryFileName = 'inventory'
 
@@ -86,8 +86,8 @@ class Ansible {
     };
     if (this.config.nodeExporter && this.config.nodeExporter.enabled) {
       data.nodeExporterEnabled = true;
-      data.nodeExporterUsername = nodeExporterUsername;
-      data.nodeExporterPassword = nodeExporterPassword;
+      data.nginxUsername = nginxUsername;
+      data.nginxPassword = nginxPassword;
       data.nodeExporterBinaryUrl = this.config.nodeExporter.binary.url;
       data.nodeExporterBinaryChecksum = this.config.nodeExporter.binary.checksum;
     } else {
