+++ src/lib/env.js
@@ -1,6 +1,6 @@
 module.exports = {
   validatorSshPrivateKeyPath: process.env.SSH_ID_RSA_VALIDATOR,
   publicNodeSshPrivateKeyPath: process.env.SSH_ID_RSA_PUBLIC,
-  nodeExporterUsername: process.env.NODE_EXPORTER_USERNAME || "prometheus",
-  nodeExporterPassword: process.env.NODE_EXPORTER_PASSWORD || "node_exporter",
+  nginxUsername: process.env.NGINX_USERNAME || "nginx_user",
+  nginxPassword: process.env.NGINX_PASSWORD || "nginx_password",
 };
