+++ tpl/ansible_inventory
@@ -24,9 +24,9 @@ polkadot_binary_checksum='{{ polkadotBinaryChecksum }}'
 chain='{{ chain }}'
 polkadot_network_id='{{ polkadotNetworkId }}'
 build_dir={{ buildDir }}
+nginx_user='{{ nginxUsername }}'
+nginx_password='{{ nginxPassword }}'
 node_exporter_enabled='{{ nodeExporterEnabled }}'
-node_exporter_user='{{ nodeExporterUsername }}'
-node_exporter_password='{{ nodeExporterPassword }}'
 node_exporter_binary_url='{{ nodeExporterBinaryUrl }}'
 node_exporter_binary_checksum='{{ nodeExporterBinaryChecksum }}'
 polkadot_restart_enabled='{{ polkadotRestartEnabled }}'
